refactor(feature:client): correct package name in path tracking view model
