+++ feature/path-tracking/src/commonMain/kotlin/com/mifos/feature/path/tracking/PathTrackingScreen.kt
@@ -48,7 +48,6 @@ import com.mifos.core.model.objects.users.UserLatLng
 import com.mifos.core.model.objects.users.UserLocation
 import com.mifos.core.ui.util.DevicePreview
 import com.mifos.feature.pathTracking.PathTrackingUiState
-import com.mifos.feature.pathTracking.PathTrackingViewModel
 import kotlinx.serialization.json.Json
 import org.jetbrains.compose.resources.stringResource
 import org.jetbrains.compose.ui.tooling.preview.PreviewParameter
